+++ .storybook/preview.ts
@@ -1,3 +1,5 @@
+import type { Preview } from "@storybook/react";
+
 export const parameters = {
   actions: { argTypesRegex: "^on[A-Z].*" },
   controls: {
@@ -7,3 +9,10 @@ export const parameters = {
     },
   },
 }
+
+const preview: Preview = {
+  parameters
+};
+
+export default preview;
+