+++ src/components/WindowsViewport.tsx
@@ -2,6 +2,7 @@ import { css, cx } from "@emotion/css";
 import { useTheme } from "@emotion/react";
 import React, { PropsWithChildren, useMemo } from "react";
 import { useSize } from "../hooks";
+import { useVisualHeight } from "../hooks/useVisualHeight";
 import { ViewportContext } from "../ViewportContext";
 
 const ignorePointerEvents = css({
@@ -20,8 +21,11 @@ const fullscreenFixed = css({
 });
 
 export function WindowsViewport({ children }: PropsWithChildren<unknown>): JSX.Element {
-  const { observe, height, width } = useSize();
-  const value = useMemo(() => ({ width, height }), [height, width]);
+  const { observe, height, width, entry } = useSize();
+  const visualHeight = useVisualHeight();
+
+  const value = useMemo(() => ({ width, height: Math.min(height, visualHeight) }), [height, width, visualHeight]);
+
   const { zIndex } = useTheme();
   return (
     <div className={cx(fullscreenFixed, ignorePointerEvents, css({ zIndex }))} ref={observe}>